+++ spec/www/spec/db-tx-error-handling-test.js
@@ -3211,6 +3211,7 @@ var mytests = function() {
         it(suiteName + 'Inline US-ASCII blank string test with arguments array=function (BOGUS) in a try-catch block [TBD (WebKit) Web SQL vs plugin]', function(done) {
           var db = openDatabase("Inline-US-ASCII-string-test-with-arguments-array-equals-function.db", "1.0", "Demo", DEFAULT_SIZE);
 
+          if (isWebSql) pending('SKIP for (WebKit) Web SQL'); // XXX TBD SKIP FOR NOW on (WebKit) Web SQL
           db.transaction(function(transaction) {
             try {
               transaction.executeSql("SELECT ''", function(ignored1, ignored2) {
@@ -3260,6 +3261,7 @@ var mytests = function() {
         it(suiteName + 'Inline US-ASCII blank string test with arguments array=function (BOGUS) in a try-catch block read tx [TBD (WebKit) Web SQL vs plugin]', function(done) {
           var db = openDatabase("Inline-US-ASCII-string-test-with-arguments-array-equals-function-read-tx.db", "1.0", "Demo", DEFAULT_SIZE);
 
+          if (isWebSql) pending('SKIP for (WebKit) Web SQL'); // XXX TBD SKIP FOR NOW on (WebKit) Web SQL
           db.readTransaction(function(readTransaction) {
             try {
               readTransaction.executeSql("SELECT ''", function(ignored1, ignored2) {
@@ -3309,6 +3311,7 @@ var mytests = function() {
         it(suiteName + 'SELECT ? test with arguments array=function (BOGUS) in a try-catch block [TBD (WebKit) Web SQL vs plugin]', function(done) {
           var db = openDatabase("SELECT-parameter-with-arguments-array-equals-function.db", "1.0", "Demo", DEFAULT_SIZE);
 
+          if (isWebSql) pending('SKIP for (WebKit) Web SQL'); // XXX TBD SKIP FOR NOW on (WebKit) Web SQL
           db.transaction(function(transaction) {
             try {
               transaction.executeSql("SELECT ?", function(ignored1, ignored2) {
@@ -3358,6 +3361,7 @@ var mytests = function() {
         it(suiteName + 'SELECT ? test with arguments array=function (BOGUS) in a try-catch block read tx [TBD (WebKit) Web SQL vs plugin]', function(done) {
           var db = openDatabase("SELECT-parameter-with-arguments-array-equals-function-read-tx.db", "1.0", "Demo", DEFAULT_SIZE);
 
+          if (isWebSql) pending('SKIP for (WebKit) Web SQL'); // XXX TBD SKIP FOR NOW on (WebKit) Web SQL
           db.readTransaction(function(transaction) {
             try {
               transaction.executeSql("SELECT ?", function(ignored1, ignored2) {