+++ src/android/io/sqlc/SQLitePlugin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2017: Christopher J. Brody (aka Chris Brody)
+ * Copyright (c) 2012-2018: Christopher J. Brody (aka Chris Brody)
  * Copyright (c) 2005-2010, Nitobi Software Inc.
  * Copyright (c) 2010, IBM Corporation
  */
@@ -11,6 +11,7 @@ import android.annotation.SuppressLint;
 import android.util.Log;
 
 import java.io.File;
+
 import java.lang.IllegalArgumentException;
 //import java.lang.Number;
 
@@ -18,7 +19,10 @@ import java.net.URI;
 import java.net.URISyntaxException;
 
 import java.util.concurrent.ConcurrentHashMap;
+
+import java.util.Map;
 import java.util.concurrent.BlockingQueue;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.LinkedBlockingQueue;
 
 import org.apache.cordova.CallbackContext;
@@ -28,11 +32,6 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-//import java.io.FileOutputStream;
-//import java.io.InputStream;
-//import java.io.OutputStream;
-//import java.io.IOException;
-
 import java.sql.SQLException;
 
 import org.apache.cordova.PluginResult;
@@ -41,11 +40,22 @@ public class SQLitePlugin extends CordovaPlugin {
 
     /**
      * Multiple database runner map (static).
-     * NOTE: no public static accessor to db (runner) map since it would not work with db threading.
-     * FUTURE put DBRunner into a public class that can provide external accessor.
+     *
+     * NOTE: no public static accessor to db (runner) map since it is not
+     * expected to work properly with db threading.
+     *
+     * FUTURE TBD put DBRunner into a public class that can provide external accessor.
+     *
+     * ADDITIONAL NOTE: Storing as Map<String, DBRunner> to avoid portabiity issue
+     * between Java 6/7/8 as discussed in:
+     * https://gist.github.com/AlainODea/1375759b8720a3f9f094
+     *
+     * THANKS to @NeoLSN (Jason Yang/楊朝傑) for giving the pointer in:
+     * https://github.com/litehelpers/Cordova-sqlite-storage/issues/727
      */
-    static ConcurrentHashMap<String, DBRunner> dbrmap = new ConcurrentHashMap<String, DBRunner>();
-    static ConcurrentHashMap<Integer, DBRunner> dbrmap2 = new ConcurrentHashMap<Integer, DBRunner>();
+    static Map<String, DBRunner> dbrmap = new ConcurrentHashMap<String, DBRunner>();
+    // static ConcurrentHashMap<Integer, DBRunner> dbrmap2 = new ConcurrentHashMap<Integer, DBRunner>();
+    static Map<Integer, DBRunner> dbrmap2 = new ConcurrentHashMap<Integer, DBRunner>();
 
     static int lastdbid = 0;
 
@@ -227,30 +237,11 @@ public class SQLitePlugin extends CordovaPlugin {
     // --------------------------------------------------------------------------
 
     private void startDatabase(String dbname, JSONObject options, CallbackContext cbc) {
-        // TODO: is it an issue that we can orphan an existing thread?  What should we do here?
-        // If we re-use the existing DBRunner it might be in the process of closing...
         DBRunner r = dbrmap.get(dbname);
 
-        // Brody TODO: It may be better to terminate the existing db thread here & start a new one, instead.
         if (r != null) {
-            // don't orphan the existing thread; just re-open the existing database.
-            // In the worst case it might be in the process of closing, but even that's less serious
-            // than orphaning the old DBRunner.
-
-            if (r.oldImpl) {
-              // Android version with 'normal' JSON interface:
-              cbc.success();
-            } else {
-              try {
-                // Android version with flat JSON interface:
-                JSONObject a1 = new JSONObject();
-                a1.put("dbid", r.dbid);
-                cbc.success(a1);
-              } catch(JSONException e) {
-                // NOT EXPECTED:
-                cbc.error("Internal error");
-              }
-            }
+            // NO LONGER EXPECTED due to BUG 666 workaround solution:
+            cbc.error("INTERNAL ERROR: database already open for db name: " + dbname);
         } else {
             r = new DBRunner(dbname, options, cbc, ++lastdbid);
             dbrmap.put(dbname, r);